requirement.txt
flask
line-bot-sdk
torch
torchvision
pillow
numpy
pandas
openpyxl
scikit-learn
